+++ src/PhpSpreadsheet/Calculation/Calculation.php
@@ -224,6 +224,16 @@ class Calculation
             'functionCall' => 'acosh',
             'argumentCount' => '1',
         ],
+        'ACOT' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'ACOT'],
+            'argumentCount' => '1',
+        ],
+        'ACOTH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'ACOTH'],
+            'argumentCount' => '1',
+        ],
         'ADDRESS' => [
             'category' => Category::CATEGORY_LOOKUP_AND_REFERENCE,
             'functionCall' => [LookupRef::class, 'cellAddress'],
@@ -425,6 +435,11 @@ class Calculation
             'functionCall' => [Engineering::class, 'COMPLEX'],
             'argumentCount' => '2,3',
         ],
+        'CONCAT' => [
+            'category' => Category::CATEGORY_TEXT_AND_DATA,
+            'functionCall' => [TextData::class, 'CONCATENATE'],
+            'argumentCount' => '1+',
+        ],
         'CONCATENATE' => [
             'category' => Category::CATEGORY_TEXT_AND_DATA,
             'functionCall' => [TextData::class, 'CONCATENATE'],
@@ -453,6 +468,16 @@ class Calculation
         'COSH' => [
             'category' => Category::CATEGORY_MATH_AND_TRIG,
             'functionCall' => 'cosh',
+            'argumentCount' => '1',
+        ],
+        'COT' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'COT'],
+            'argumentCount' => '1',
+        ],
+        'COTH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'COTH'],
             'argumentCount' => '1',
         ],
         'COUNT' => [
@@ -519,6 +544,16 @@ class Calculation
             'category' => Category::CATEGORY_STATISTICAL,
             'functionCall' => [Statistical::class, 'CRITBINOM'],
             'argumentCount' => '3',
+        ],
+        'CSC' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'CSC'],
+            'argumentCount' => '1',
+        ],
+        'CSCH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'CSCH'],
+            'argumentCount' => '1',
         ],
         'CUBEKPIMEMBER' => [
             'category' => Category::CATEGORY_CUBE,
@@ -735,7 +770,17 @@ class Calculation
             'functionCall' => [Engineering::class, 'ERF'],
             'argumentCount' => '1,2',
         ],
+        'ERF.PRECISE' => [
+            'category' => Category::CATEGORY_ENGINEERING,
+            'functionCall' => [Engineering::class, 'ERFPRECISE'],
+            'argumentCount' => '1',
+        ],
         'ERFC' => [
+            'category' => Category::CATEGORY_ENGINEERING,
+            'functionCall' => [Engineering::class, 'ERFC'],
+            'argumentCount' => '1',
+        ],
+        'ERFC.PRECISE' => [
             'category' => Category::CATEGORY_ENGINEERING,
             'functionCall' => [Engineering::class, 'ERFC'],
             'argumentCount' => '1',
@@ -824,6 +869,13 @@ class Calculation
             'category' => Category::CATEGORY_STATISTICAL,
             'functionCall' => [Statistical::class, 'FORECAST'],
             'argumentCount' => '3',
+        ],
+        'FORMULATEXT' => [
+            'category' => Category::CATEGORY_LOOKUP_AND_REFERENCE,
+            'functionCall' => [LookupRef::class, 'FORMULATEXT'],
+            'argumentCount' => '1',
+            'passCellReference' => true,
+            'passByReference' => [true],
         ],
         'FREQUENCY' => [
             'category' => Category::CATEGORY_STATISTICAL,
@@ -1114,6 +1166,11 @@ class Calculation
             'functionCall' => [Functions::class, 'isOdd'],
             'argumentCount' => '1',
         ],
+        'ISOWEEKNUM' => [
+            'category' => Category::CATEGORY_DATE_AND_TIME,
+            'functionCall' => [DateTime::class, 'ISOWEEKNUM'],
+            'argumentCount' => '1',
+        ],
         'ISPMT' => [
             'category' => Category::CATEGORY_FINANCIAL,
             'functionCall' => [Financial::class, 'ISPMT'],
@@ -1394,6 +1451,11 @@ class Calculation
             'functionCall' => [Financial::class, 'NPV'],
             'argumentCount' => '2+',
         ],
+        'NUMBERVALUE' => [
+            'category' => Category::CATEGORY_TEXT_AND_DATA,
+            'functionCall' => [TextData::class, 'NUMBERVALUE'],
+            'argumentCount' => '1+',
+        ],
         'OCT2BIN' => [
             'category' => Category::CATEGORY_ENGINEERING,
             'functionCall' => [Engineering::class, 'OCTTOBIN'],
@@ -1445,6 +1507,11 @@ class Calculation
             'category' => Category::CATEGORY_LOGICAL,
             'functionCall' => [Logical::class, 'logicalOr'],
             'argumentCount' => '1+',
+        ],
+        'PDURATION' => [
+            'category' => Category::CATEGORY_FINANCIAL,
+            'functionCall' => [Financial::class, 'PDURATION'],
+            'argumentCount' => '3',
         ],
         'PEARSON' => [
             'category' => Category::CATEGORY_STATISTICAL,
@@ -1627,6 +1694,11 @@ class Calculation
             'functionCall' => [LookupRef::class, 'ROWS'],
             'argumentCount' => '1',
         ],
+        'RRI' => [
+            'category' => Category::CATEGORY_FINANCIAL,
+            'functionCall' => [Financial::class, 'RRI'],
+            'argumentCount' => '3',
+        ],
         'RSQ' => [
             'category' => Category::CATEGORY_STATISTICAL,
             'functionCall' => [Statistical::class, 'RSQ'],
@@ -1646,6 +1718,16 @@ class Calculation
             'category' => Category::CATEGORY_TEXT_AND_DATA,
             'functionCall' => [TextData::class, 'SEARCHINSENSITIVE'],
             'argumentCount' => '2,3',
+        ],
+        'SEC' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'SEC'],
+            'argumentCount' => '1',
+        ],
+        'SECH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'SECH'],
+            'argumentCount' => '1',
         ],
         'SECOND' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
@@ -1838,6 +1920,11 @@ class Calculation
             'functionCall' => [TextData::class, 'TEXTFORMAT'],
             'argumentCount' => '2',
         ],
+        'TEXTJOIN' => [
+            'category' => Category::CATEGORY_TEXT_AND_DATA,
+            'functionCall' => [TextData::class, 'TEXTJOIN'],
+            'argumentCount' => '3+',
+        ],
         'TIME' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
             'functionCall' => [DateTime::class, 'TIME'],
@@ -1898,6 +1985,16 @@ class Calculation
             'functionCall' => [Functions::class, 'TYPE'],
             'argumentCount' => '1',
         ],
+        'UNICHAR' => [
+            'category' => Category::CATEGORY_TEXT_AND_DATA,
+            'functionCall' => [TextData::class, 'CHARACTER'],
+            'argumentCount' => '1',
+        ],
+        'UNICODE' => [
+            'category' => Category::CATEGORY_TEXT_AND_DATA,
+            'functionCall' => [TextData::class, 'ASCIICODE'],
+            'argumentCount' => '1',
+        ],
         'UPPER' => [
             'category' => Category::CATEGORY_TEXT_AND_DATA,
             'functionCall' => [TextData::class, 'UPPERCASE'],
@@ -1914,6 +2011,16 @@ class Calculation
             'argumentCount' => '1',
         ],
         'VAR' => [
+            'category' => Category::CATEGORY_STATISTICAL,
+            'functionCall' => [Statistical::class, 'VARFunc'],
+            'argumentCount' => '1+',
+        ],
+        'VAR.P' => [
+            'category' => Category::CATEGORY_STATISTICAL,
+            'functionCall' => [Statistical::class, 'VARP'],
+            'argumentCount' => '1+',
+        ],
+        'VAR.S' => [
             'category' => Category::CATEGORY_STATISTICAL,
             'functionCall' => [Statistical::class, 'VARFunc'],
             'argumentCount' => '1+',
@@ -1972,6 +2079,11 @@ class Calculation
             'category' => Category::CATEGORY_FINANCIAL,
             'functionCall' => [Financial::class, 'XNPV'],
             'argumentCount' => '3',
+        ],
+        'XOR' => [
+            'category' => Category::CATEGORY_LOGICAL,
+            'functionCall' => [Logical::class, 'logicalXor'],
+            'argumentCount' => '1+',
         ],
         'YEAR' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,