+++ tests/PhpSpreadsheetTests/Calculation/FunctionsTest.php
@@ -4,6 +4,7 @@ namespace PhpOffice\PhpSpreadsheetTests\Calculation;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Functions;
 use PhpOffice\PhpSpreadsheet\Cell\Cell;
+use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 use PHPUnit\Framework\TestCase;
 
@@ -12,6 +13,49 @@ class FunctionsTest extends TestCase
     public function setUp()
     {
         Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        Functions::setReturnDateType(Functions::RETURNDATE_EXCEL);
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        Functions::setReturnDateType(Functions::RETURNDATE_EXCEL);
+    }
+
+    public function testCompatibilityMode()
+    {
+        $result = Functions::setCompatibilityMode(Functions::COMPATIBILITY_GNUMERIC);
+        // Test for a true response for success
+        $this->assertTrue($result);
+        // Test that mode has been changed
+        $this->assertEquals(Functions::COMPATIBILITY_GNUMERIC, Functions::getCompatibilityMode());
+    }
+
+    public function testInvalidCompatibilityMode()
+    {
+        $result = Functions::setCompatibilityMode('INVALIDMODE');
+        // Test for a false response for failure
+        $this->assertFalse($result);
+        // Test that mode has not been changed
+        $this->assertEquals(Functions::COMPATIBILITY_EXCEL, Functions::getCompatibilityMode());
+    }
+
+    public function testReturnDateType()
+    {
+        $result = Functions::setReturnDateType(Functions::RETURNDATE_PHP_OBJECT);
+        // Test for a true response for success
+        $this->assertTrue($result);
+        // Test that mode has been changed
+        $this->assertEquals(Functions::RETURNDATE_PHP_OBJECT, Functions::getReturnDateType());
+    }
+
+    public function testInvalidReturnDateType()
+    {
+        $result = Functions::setReturnDateType('INVALIDTYPE');
+        // Test for a false response for failure
+        $this->assertFalse($result);
+        // Test that mode has not been changed
+        $this->assertEquals(Functions::RETURNDATE_EXCEL, Functions::getReturnDateType());
     }
 
     public function testDUMMY()
@@ -274,23 +318,38 @@ class FunctionsTest extends TestCase
      * @dataProvider providerIsFormula
      *
      * @param mixed $expectedResult
-     * @param mixed $value
+     * @param mixed $reference       Reference to the cell we wish to test
+     * @param mixed $value           Value of the cell we wish to test
      */
-    public function testIsFormula($expectedResult, $value = 'undefined')
+    public function testIsFormula($expectedResult, $reference, $value = 'undefined')
     {
         $ourCell = null;
         if ($value !== 'undefined') {
             $remoteCell = $this->getMockBuilder(Cell::class)
                 ->disableOriginalConstructor()
                 ->getMock();
-            $remoteCell->method('getValue')
-                ->will($this->returnValue($value));
+            $remoteCell->method('isFormula')
+                ->will($this->returnValue(substr($value, 0, 1) == '='));
+
+            $remoteSheet = $this->getMockBuilder(Worksheet::class)
+                ->disableOriginalConstructor()
+                ->getMock();
+            $remoteSheet->method('getCell')
+                ->will($this->returnValue($remoteCell));
+
+            $workbook = $this->getMockBuilder(Spreadsheet::class)
+                ->disableOriginalConstructor()
+                ->getMock();
+            $workbook->method('getSheetByName')
+                ->will($this->returnValue($remoteSheet));
 
             $sheet = $this->getMockBuilder(Worksheet::class)
                 ->disableOriginalConstructor()
                 ->getMock();
             $sheet->method('getCell')
                 ->will($this->returnValue($remoteCell));
+            $sheet->method('getParent')
+                ->will($this->returnValue($workbook));
 
             $ourCell = $this->getMockBuilder(Cell::class)
                 ->disableOriginalConstructor()
@@ -299,7 +358,7 @@ class FunctionsTest extends TestCase
                 ->will($this->returnValue($sheet));
         }
 
-        $result = Functions::isFormula($value, $ourCell);
+        $result = Functions::isFormula($reference, $ourCell);
         self::assertEquals($expectedResult, $result, null, 1E-8);
     }